Move "StreamChunkReader" to own file
