+++ tests/Tests.cs
@@ -213,42 +213,3 @@ public class PartialReadTests
         }
     }
 }
-
-public sealed class StreamChunkReader : IDisposable
-{
-    readonly Stream stream;
-    byte[] buffer;
-    ReadOnlyMemory<byte> memory;
-
-    public StreamChunkReader(Stream stream, int bufferSize)
-    {
-        this.stream = stream;
-        this.buffer = new byte[bufferSize];
-        this.memory = null;
-    }
-
-    public async ValueTask<ReadOnlyMemory<byte>> ReadAsync(int bytesConsumed, CancellationToken cancellationToken)
-    {
-        int bytesRead;
-
-        var restLength = this.memory.Length - bytesConsumed;
-        if (restLength > 0)
-        {
-            ReadOnlyMemory<byte> rest = this.buffer.AsMemory(^restLength..);
-
-            if (rest.Length == this.buffer.Length)
-                Array.Resize(ref this.buffer, this.buffer.Length * 2);
-
-            rest.CopyTo(this.buffer);
-            bytesRead = await this.stream.ReadAsync(this.buffer.AsMemory(rest.Length), cancellationToken).ConfigureAwait(false);
-        }
-        else
-        {
-            bytesRead = await this.stream.ReadAsync(this.buffer, cancellationToken).ConfigureAwait(false);
-        }
-
-        return this.memory = this.buffer.AsMemory(..(restLength + bytesRead));
-    }
-
-    public void Dispose() => this.stream.Dispose();
-}
